refactor NoPageAlert
