+++ src/components/domain/Page/molecules/PageList/index.tsx
@@ -4,7 +4,7 @@ import { Page } from '~/domains/Page';
 import { useOgpCardLayout } from '~/stores/contexts';
 import { OgpLayoutType } from '~/interfaces/contexts';
 
-import { NoPageAlert } from '~/components/Alerts/NoPageAlert';
+import { NoPageAlert } from '~/components/domain/Page/molecules/NoPageAlert';
 import { PaginationWrapper } from '~/components/Commons/PaginationWrapper';
 import { PageListItem } from '~/components/domain/Page/molecules/PageListItem';
 import { PageCard } from '~/components/domain/Page/molecules/PageCard';
